+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add new `RSpec/Eq` cop. ([@ydah])
 - Fix an incorrect autocorrect for `RSpec/ReceiveMessages` when return values declared between stubs. ([@marocchino])
 - Split `RSpec/FilePath` into `RSpec/SpecFilePathSuffix` and `RSpec/SpecFilePathFormat`. `RSpec/FilePath` cop is enabled by default, the two new cups are pending and need to be enabled explicitly. ([@ydah])
+- Add support `RSpec/Rails/HttpStatus` when `have_http_status` with string argument. ([@ydah])
 
 ## 2.23.2 (2023-08-09)